Made a good dent with the warnings.
