+++ DFC.App.CareerPath.IntegrationTests.API/Model/CareerPathAPIResponse.cs
@@ -1,9 +1,11 @@
-﻿using System.Collections.Generic;
+﻿using Newtonsoft.Json;
+using System.Collections.Generic;
 
 namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Model
 {
     internal class CareerPathAPIResponse
     {
-        public List<string> careerPathAndProgression { get; set; }
+        [JsonProperty("careerPathAndProgression")]
+        public List<string> CareerPathAndProgression { get; set; }
     }
 }
