+++ DFC.App.CareerPath.IntegrationTests.API/Model/JobProfile/Environment.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Model.JobProfile
+{
+    public class Environment
+    {
+        public string Id { get; set; }
+
+        public string Title { get; set; }
+
+        public string Description { get; set; }
+
+        public string Url { get; set; }
+
+        public bool IsNegative { get; set; }
+    }
+}
