+++ DFC.App.CareerPath.IntegrationTests.API/Model/JobProfile/FurtherInformationContentType.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Model.JobProfile
+{
+    public class FurtherInformationContentType
+    {
+        public string CareerTips { get; set; }
+
+        public string ProfessionalAndIndustryBodies { get; set; }
+
+        public string FurtherInformation { get; set; }
+    }
+}
