+++ DFC.App.CareerPath.IntegrationTests.API/Model/JobProfile/FurtherRoutes.cs
@@ -1,0 +1,13 @@
+﻿namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Model.JobProfile
+{
+    public class FurtherRoutes
+    {
+        public string Work { get; set; }
+
+        public string Volunteering { get; set; }
+
+        public string DirectApplication { get; set; }
+
+        public string OtherRoutes { get; set; }
+    }
+}
