+++ DFC.App.CareerPath.IntegrationTests.API/Model/JobProfile/HiddenAlternativeTitle.cs
@@ -1,0 +1,13 @@
+﻿namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Model.JobProfile
+{
+    public class HiddenAlternativeTitle
+    {
+        public string Id { get; set; }
+
+        public string Title { get; set; }
+
+        public string Url { get; set; }
+
+        public string Description { get; set; }
+    }
+}
