+++ DFC.App.CareerPath.IntegrationTests.API/Model/JobProfile/JobProfileCategory.cs
@@ -1,0 +1,11 @@
+﻿namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Model.JobProfile
+{
+    public class JobProfileCategory
+    {
+        public string Id { get; set; }
+
+        public string Name { get; set; }
+
+        public string Title { get; set; }
+    }
+}
