+++ DFC.App.CareerPath.IntegrationTests.API/Model/JobProfile/MoreInformationLink.cs
@@ -1,0 +1,13 @@
+﻿namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Model.JobProfile
+{
+    public class MoreInformationLink
+    {
+        public string Id { get; set; }
+
+        public string Title { get; set; }
+
+        public string Url { get; set; }
+
+        public string Text { get; set; }
+    }
+}
