+++ DFC.App.CareerPath.IntegrationTests.API/Model/JobProfile/Registration.cs
@@ -1,0 +1,11 @@
+﻿namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Model.JobProfile
+{
+    public class Registration
+    {
+        public string Id { get; set; }
+
+        public string Title { get; set; }
+
+        public string Info { get; set; }
+    }
+}
