+++ DFC.App.CareerPath.IntegrationTests.API/Model/JobProfile/RelatedCareersData.cs
@@ -1,0 +1,11 @@
+﻿namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Model.JobProfile
+{
+    public class RelatedCareersData
+    {
+        public string Id { get; set; }
+
+        public string Title { get; set; }
+
+        public string ProfileLink { get; set; }
+    }
+}
