+++ DFC.App.CareerPath.IntegrationTests.API/Model/JobProfile/RelatedSOC.cs
@@ -1,0 +1,9 @@
+﻿namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Model.JobProfile
+{
+    public class RelatedSOC
+    {
+        public string Id { get; set; }
+
+        public string SOCCode { get; set; }
+    }
+}
