+++ DFC.App.CareerPath.IntegrationTests.API/Model/JobProfile/RelatedSkill.cs
@@ -1,0 +1,13 @@
+﻿namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Model.JobProfile
+{
+    public class RelatedSkill
+    {
+        public string Id { get; set; }
+
+        public string Title { get; set; }
+
+        public string ONetElementId { get; set; }
+
+        public string Description { get; set; }
+    }
+}
