+++ DFC.App.CareerPath.IntegrationTests.API/Model/JobProfile/Restriction.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Model.JobProfile
+{
+    public class Restriction
+    {
+        public string Id { get; set; }
+
+        public string Title { get; set; }
+
+        public string Info { get; set; }
+    }
+}
