+++ DFC.App.CareerPath.IntegrationTests.API/Model/JobProfile/RouteEntry.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Model.JobProfile
+{
+    public class RouteEntry
+    {
+        public int RouteName { get; set; }
+
+        public List<EntryRequirement> EntryRequirements { get; set; }
+
+        public List<MoreInformationLink> MoreInformationLinks { get; set; }
+
+        public string RouteSubjects { get; set; }
+
+        public string FurtherRouteInformation { get; set; }
+
+        public string RouteRequirement { get; set; }
+    }
+}
