+++ DFC.App.CareerPath.IntegrationTests.API/Model/JobProfile/SocSkillsMatrixData.cs
@@ -1,0 +1,23 @@
+﻿using System.Collections.Generic;
+
+namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Model.JobProfile
+{
+    public class SocSkillsMatrixData
+    {
+        public string Id { get; set; }
+
+        public string Title { get; set; }
+
+        public string Contextualised { get; set; }
+
+        public string ONetAttributeType { get; set; }
+
+        public double Rank { get; set; }
+
+        public double ONetRank { get; set; }
+
+        public List<RelatedSkill> RelatedSkill { get; set; }
+
+        public List<RelatedSOC> RelatedSOC { get; set; }
+    }
+}
