+++ DFC.App.CareerPath.IntegrationTests.API/Model/JobProfile/Uniform.cs
@@ -1,0 +1,15 @@
+﻿namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Model.JobProfile
+{
+    public class Uniform
+    {
+        public string Id { get; set; }
+
+        public string Title { get; set; }
+
+        public string Description { get; set; }
+
+        public string Url { get; set; }
+
+        public bool IsNegative { get; set; }
+    }
+}
