+++ DFC.App.CareerPath.IntegrationTests.API/Model/JobProfile/WhatYouWillDoData.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections.Generic;
+
+namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Model.JobProfile
+{
+    public class WhatYouWillDoData
+    {
+        public string DailyTasks { get; set; }
+
+        public List<Location> Locations { get; set; }
+
+        public List<Uniform> Uniforms { get; set; }
+
+        public List<Environment> Environments { get; set; }
+
+        public string Introduction { get; set; }
+    }
+}
