+++ DFC.App.CareerPath.IntegrationTests.API/Model/JobProfile/WorkingHoursDetail.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Model.JobProfile
+{
+    public class WorkingHoursDetail
+    {
+        public string Id { get; set; }
+
+        public string Title { get; set; }
+
+        public string Url { get; set; }
+
+        public string Description { get; set; }
+    }
+}
