+++ DFC.App.CareerPath.IntegrationTests.API/Support/Interface/IAPISupport.cs
@@ -4,8 +4,8 @@ using static DFC.App.RelatedCareers.Tests.Common.APISupport.GetRequest;
 
 namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Support.Interface
 {
-    interface IAPISupport
+    internal interface IAPISupport
     {
-        Task<Response<T>> ExecuteGetRequest<T>(string endpoint, ContentType responseFormat, bool AuthoriseRequest = true);
+        Task<Response<T>> ExecuteGetRequest<T>(string endpoint, ContentType responseFormat, bool authoriseRequest = true);
     }
 }
