+++ DFC.App.CareerPath.IntegrationTests.API/Support/Interface/IGeneralSupport.cs
@@ -5,8 +5,11 @@ namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Support.Interface
     internal interface IGeneralSupport
     {
         string RandomString(int length);
+
         void InitialiseAppSettings();
+
         byte[] ConvertObjectToByteArray(object obj);
+
         string GetDescription(Enum enumerator);
     }
 }
