+++ DFC.App.CareerPath.IntegrationTests.API/Support/Interface/IJobProfileSupport.cs
@@ -4,9 +4,10 @@ using System.Threading.Tasks;
 
 namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Support.Interface
 {
-    interface IJobProfileSupport
+    internal interface IJobProfileSupport
     {
         JobProfileContentType GenerateJobProfileContentType();
+
         Task DeleteJobProfile(Topic topic, JobProfileContentType jobProfileId);
     }
 }
