+++ DFC.App.CareerPath.IntegrationTests.API/Support/Interface/IRelatedCareersSupport.cs
@@ -1,8 +1,9 @@
 ﻿using DFC.App.RelatedCareers.Tests.IntegrationTests.API.Model;
+using DFC.App.RelatedCareers.Tests.IntegrationTests.API.Model.JobProfile;
 
 namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Support.Interface
 {
-    interface IRelatedCareersSupport
+    internal interface IRelatedCareersSupport
     {
         RelatedCareersData GenerateRelatedCareersDataSection();
     }
