+++ DFC.App.CareerPath.IntegrationTests.API/Support/ResourceManager.cs
@@ -1,27 +1,34 @@
 ﻿using Newtonsoft.Json;
 using System;
+using System.Globalization;
 using System.IO;
 
 namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Support
 {
     internal class ResourceManager
     {
+        internal static T GetResource<T>(string resourceName)
+        {
+            string content = GetResourceContent(resourceName);
+            return JsonConvert.DeserializeObject<T>(content);
+        }
+
         private static string GetResourceContent(string resourceName)
         {
             DirectoryInfo resourcesDirectory = Directory.CreateDirectory(Environment.CurrentDirectory).GetDirectories("Resource")[0];
             FileInfo[] files = resourcesDirectory.GetFiles();
             FileInfo selectedResource = null;
 
-            for(int fileIndex = 0; fileIndex < files.Length; fileIndex++)
+            for (int fileIndex = 0; fileIndex < files.Length; fileIndex++)
             {
-                if(files[fileIndex].Name.ToLower().StartsWith(resourceName.ToLower()))
+                if (files[fileIndex].Name.StartsWith(resourceName, StringComparison.CurrentCultureIgnoreCase))
                 {
                     selectedResource = files[fileIndex];
                     break;
                 }
             }
 
-            if(selectedResource.FullName == null)
+            if (selectedResource.FullName == null)
             {
                 throw new Exception($"No resource with the name {resourceName} was found");
             }
@@ -30,12 +37,6 @@ namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Support
             {
                 return streamReader.ReadToEnd();
             }
-        }
-
-        internal static T GetResource<T>(string resourceName)
-        {
-            string content = GetResourceContent(resourceName);
-            return JsonConvert.DeserializeObject<T>(content);
         }
     }
 }
